+++ src/booklab/__init__.py
@@ -22,10 +22,9 @@ PACKAGE_ROOT = pathlib.Path(__file__).parent.resolve()
 DATA_ROOT = PACKAGE_ROOT.joinpath("data")  # database files directory
 CONF_ROOT = PACKAGE_ROOT.joinpath("conf")  # config files directory
 
-# test & debug print ... TO BE REMOVED @ RELEASE
-print("*** PACKAGE_ROOT dir: ", PACKAGE_ROOT)
-print("*** DATA_ROOT dir: ", DATA_ROOT)
-print("*** CONF_ROIT dir: ", CONF_ROOT)
+# 4dbg... print("*** PACKAGE_ROOT dir: ", PACKAGE_ROOT)
+# 4dbg... print("*** DATA_ROOT dir: ", DATA_ROOT)
+# 4dbg... print("*** CONF_ROIT dir: ", CONF_ROOT)