+++ src/booklab/booklabd/db_init.py
@@ -12,22 +12,16 @@ from booklab import DATA_ROOT
 
 def init_db() -> tuple[pysondb]:
     """Create db_books, db_system objects and return them as tuple
-    
+
     Return:
 
-        (db_books, db_system) tuple `pysondb` object
+        `(db_books, db_system)` tuple of pysondb object
     """
+    file = os.path.join(DATA_ROOT, "books_catalog.json")
+    db_books = pysondb.db.getDb(file)
 
-    # ...tbd review and update code ...
-    bks_catalog_file = os.path.join(DATA_ROOT, "books_catalog.json")
-    db_books = pysondb.db.getDb(bks_catalog_file)
-
-    # bks_sysinfo_fioe = ...
-    db_system = None  # ...#TODO tbd@250831
-    #  the other db is app_info.json 4dbg... 
-    # print(f"*** booklabd.__init__ imported 
-    # {PACKAGE_ROOT=} {DATA_ROOT=} 
-    # {CONF_ROOT=}")
+    file = os.path.join(DATA_ROOT, "app_info.json")
+    db_system = pysondb.db.getDb(file)
 
     return (db_books, db_system)