+++ CHANGELOG.md
@@ -40,10 +40,10 @@ First booklabd as Flask web app to serve `/api/.../` routes
 Create database (aka `db*`) objects for JSON files and a test `/api/bcat` route:
 
 * ...tbd a test route for /api/bcat/ to write sone data in books catalog
-* ...wip 250831-c init the `db` objects to JSON databases:
+* 250831-c init the `db` objects to JSON databases:
     * `init_db.py` module and ref it in `__init__.py`
     * `db_books` for booklab.DATA_ROOT --> books_catalog.json
-    * ...tbd booklab.DATA_ROOT --> app_info.json
+    * `db_system` for booklab.DATA_ROOT --> app_info.json
 * 250831-b `booklab.booklabd.__init__` imported working directories (created in prev change)
 * 250831-a `booklab.__init__` create constants: PACKAGE _ROOT, DATA_ROOT, CONF_ROOT
 * 250830-b make gunicorn config python module in `booklab.conf` module (as new one)